Created Sentiment.kt enumeration for different sentiments, along with a SentimentCard.kt for displaying sentiments. Refactored TrendingStock.kt to hold a Sentiment object rather than a string. And created TrendingStockEntity.kt for holding a trending stock in the database. Mapping between TrendingStock, TrendingStockDto, and TrendingStockEntity are now done through companion objects, wrather than a mapper file.
Sentiment for a stock is now passed along with the ticker to navigation destination in TrendiesNavHost.kt, for easy passing, eliminating a resource intensive call to fetch an element from the database for one piece of information.

Still need to implement drawing of SentimentCard.kt on both the list and detail screens, along with updating navigation on both screens to pass and receive sentiment along with the ticker.